cv_generator/requirements.txt
streamlit
jinja2
google-generativeai
pymupdf
